# Nest.js/3/demo/swagger.json
{"openapi":"3.0.0","paths":{"/tasks":{"get":{"operationId":"TasksController_getAllTasks","parameters":[],"responses":{"200":{"description":""}},"tags":["tasks"]},"post":{"operationId":"TasksController_createTask","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateTaskDto"}}}},"responses":{"201":{"description":""}},"tags":["tasks"]}},"/tasks/{id}":{"get":{"operationId":"TasksController_getTaskById","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TaskVM"}}}}},"tags":["tasks"]},"delete":{"operationId":"TasksController_deleteTaskById","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"204":{"description":""}},"tags":["tasks"]},"put":{"operationId":"TasksController_updateTask","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateTaskDto"}}}},"responses":{"200":{"description":""}},"tags":["tasks"]}},"/auth/signup":{"post":{"operationId":"AuthController_signup","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthCredentialsDto"}}}},"responses":{"201":{"description":""}},"tags":["Auth"]}},"/auth/signin":{"post":{"operationId":"AuthController_signIn","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginVM"}}}},"401":{"description":"","content":{"application/json":{"schema":{"type":"string"}}}}},"tags":["Auth"]}},"/auth/test":{"post":{"operationId":"AuthController_test","parameters":[],"responses":{"201":{"description":""}},"tags":["Auth"]}}},"info":{"title":"Your API Title","description":"API description","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"schemas":{"TaskVM":{"type":"object","properties":{"id":{"type":"string"},"title":{"type":"string"},"description":{"type":"string"},"status":{"type":"string"}},"required":["id","title","description","status"]},"CreateTaskDto":{"type":"object","properties":{"title":{"type":"string"},"description":{"type":"string"}},"required":["title","description"]},"UpdateTaskDto":{"type":"object","properties":{"title":{"type":"string"},"description":{"type":"string"},"status":{"type":"string"}},"required":["title","description","status"]},"AuthCredentialsDto":{"type":"object","properties":{"username":{"type":"string"},"email":{"type":"string"},"password":{"type":"string"}},"required":["username","email","password"]},"LoginDto":{"type":"object","properties":{"username":{"type":"string"},"password":{"type":"string"}},"required":["username","password"]},"LoginVM":{"type":"object","properties":{"accessToken":{"type":"string"}},"required":["accessToken"]}}}}